+++ app/src/main/java/net/kullo/android/application/CacheType.java
@@ -6,4 +6,5 @@ public enum CacheType {
     Capture,
     OpenFile,
     ReceivedShares,
+    ReceivedSharePreviews,
 }