+++ app/src/main/java/net/kullo/android/screens/ConversationsListActivity.java
@@ -346,7 +346,7 @@ public class ConversationsListActivity extends AppCompatActivity implements
                         mSwipeLayout.setRefreshing(false);
 
                         if (KulloUtils.showSyncProgressAsBar(progress)) {
-                            int percent = Math.round(100 * ((float) progress.getCountProcessed() / progress.getCountTotal()));
+                            int percent = Math.round(100 * ((float) progress.getIncomingMessagesProcessed() / progress.getIncomingMessagesTotal()));
 
                             mProgressBarIndeterminate.setVisibility(View.GONE);
                             mProgressBarDeterminate.setVisibility(View.VISIBLE);