+++ app/src/main/java/net/kullo/android/screens/sharereceiver/ImagesPreviewAdapter.java
@@ -1,10 +1,13 @@
 /* Copyright 2015-2016 Kullo GmbH. All rights reserved. */
 package net.kullo.android.screens.sharereceiver;
 
+import android.content.Context;
 import android.support.v7.widget.RecyclerView;
 import android.view.LayoutInflater;
 import android.view.View;
 import android.view.ViewGroup;
+
+import com.bumptech.glide.Glide;
 
 import net.kullo.android.R;
 import net.kullo.android.littlehelpers.Formatting;
@@ -13,7 +16,12 @@ import java.util.ArrayList;
 
 public class ImagesPreviewAdapter extends RecyclerView.Adapter<ImagesPreviewViewHolder> {
 
+    private final Context mContext;
     private ArrayList<Share> mShares = new ArrayList<>();
+
+    public ImagesPreviewAdapter(Context context) {
+        mContext = context;
+    }
 
     @Override
     public ImagesPreviewViewHolder onCreateViewHolder(ViewGroup parent, int viewType) {
@@ -25,10 +33,10 @@ public class ImagesPreviewAdapter extends RecyclerView.Adapter<ImagesPreviewView
     @Override
     public void onBindViewHolder(ImagesPreviewViewHolder holder, int position) {
         Share share = mShares.get(position);
-        holder.mImage.setImageBitmap(share.preview);
         String metaText = String.format("%s (%s)",
                 share.filename, Formatting.filesizeHuman(share.size));
         holder.mImageMetaText.setText(metaText);
+        Glide.with(mContext).load(share.previewUri).into(holder.mImage);
     }
 
     @Override