+++ libkullo/src/main/java/net/kullo/libkullo/api/DraftAttachmentsAddListener.java
@@ -5,6 +5,12 @@ package net.kullo.libkullo.api;
 
 /** Listener used in DraftAttachments.addAsync() */
 public abstract class DraftAttachmentsAddListener {
+    /**
+     * Indicates progress when adding an attachment.
+     * Note: not guaranteed to be called at any specific point.
+     */
+    public abstract void progressed(long convId, long attId, long bytesProcessed, long bytesTotal);
+
     public abstract void finished(long convId, long attId, String path);
 
     public abstract void error(long convId, String path, LocalError error);