+++ libkullo/src/main/java/net/kullo/libkullo/api/SyncProgress.java
@@ -3,74 +3,101 @@
 
 package net.kullo.libkullo.api;
 
+/** Encodes a progress update during syncing. Unknown totals are set to 0. */
 public final class SyncProgress {
 
 
-    /*package*/ final long countLeft;
+    /*package*/ final long incomingMessagesProcessed;
 
-    /*package*/ final long countProcessed;
+    /*package*/ final long incomingMessagesTotal;
 
-    /*package*/ final long countTotal;
+    /*package*/ final long incomingMessagesNew;
 
-    /*package*/ final long countNew;
+    /*package*/ final long incomingMessagesNewUnread;
 
-    /*package*/ final long countNewUnread;
+    /*package*/ final long incomingMessagesModified;
 
-    /*package*/ final long countModified;
+    /*package*/ final long incomingMessagesDeleted;
 
-    /*package*/ final long countDeleted;
+    /*package*/ final long incomingAttachmentsDownloadedBytes;
+
+    /*package*/ final long incomingAttachmentsTotalBytes;
+
+    /*package*/ final long outgoingMessagesUploadedBytes;
+
+    /*package*/ final long outgoingMessagesTotalBytes;
 
     /*package*/ final long runTimeMs;
 
     public SyncProgress(
-            long countLeft,
-            long countProcessed,
-            long countTotal,
-            long countNew,
-            long countNewUnread,
-            long countModified,
-            long countDeleted,
+            long incomingMessagesProcessed,
+            long incomingMessagesTotal,
+            long incomingMessagesNew,
+            long incomingMessagesNewUnread,
+            long incomingMessagesModified,
+            long incomingMessagesDeleted,
+            long incomingAttachmentsDownloadedBytes,
+            long incomingAttachmentsTotalBytes,
+            long outgoingMessagesUploadedBytes,
+            long outgoingMessagesTotalBytes,
             long runTimeMs) {
-        this.countLeft = countLeft;
-        this.countProcessed = countProcessed;
-        this.countTotal = countTotal;
-        this.countNew = countNew;
-        this.countNewUnread = countNewUnread;
-        this.countModified = countModified;
-        this.countDeleted = countDeleted;
+        this.incomingMessagesProcessed = incomingMessagesProcessed;
+        this.incomingMessagesTotal = incomingMessagesTotal;
+        this.incomingMessagesNew = incomingMessagesNew;
+        this.incomingMessagesNewUnread = incomingMessagesNewUnread;
+        this.incomingMessagesModified = incomingMessagesModified;
+        this.incomingMessagesDeleted = incomingMessagesDeleted;
+        this.incomingAttachmentsDownloadedBytes = incomingAttachmentsDownloadedBytes;
+        this.incomingAttachmentsTotalBytes = incomingAttachmentsTotalBytes;
+        this.outgoingMessagesUploadedBytes = outgoingMessagesUploadedBytes;
+        this.outgoingMessagesTotalBytes = outgoingMessagesTotalBytes;
         this.runTimeMs = runTimeMs;
     }
 
-    /** stats about messages */
-    public long getCountLeft() {
-        return countLeft;
+    /** inbox (unit: messages) */
+    public long getIncomingMessagesProcessed() {
+        return incomingMessagesProcessed;
     }
 
-    public long getCountProcessed() {
-        return countProcessed;
+    public long getIncomingMessagesTotal() {
+        return incomingMessagesTotal;
     }
 
-    public long getCountTotal() {
-        return countTotal;
+    public long getIncomingMessagesNew() {
+        return incomingMessagesNew;
     }
 
-    public long getCountNew() {
-        return countNew;
+    public long getIncomingMessagesNewUnread() {
+        return incomingMessagesNewUnread;
     }
 
-    public long getCountNewUnread() {
-        return countNewUnread;
+    public long getIncomingMessagesModified() {
+        return incomingMessagesModified;
     }
 
-    public long getCountModified() {
-        return countModified;
+    public long getIncomingMessagesDeleted() {
+        return incomingMessagesDeleted;
     }
 
-    public long getCountDeleted() {
-        return countDeleted;
+    /** incoming attachments (unit: bytes) */
+    public long getIncomingAttachmentsDownloadedBytes() {
+        return incomingAttachmentsDownloadedBytes;
     }
 
-    /** Run time of the current sync in milliseconds */
+    public long getIncomingAttachmentsTotalBytes() {
+        return incomingAttachmentsTotalBytes;
+    }
+
+    /** outgoing messages + attachments (unit: uncompressed bytes) */
+    public long getOutgoingMessagesUploadedBytes() {
+        return outgoingMessagesUploadedBytes;
+    }
+
+    public long getOutgoingMessagesTotalBytes() {
+        return outgoingMessagesTotalBytes;
+    }
+
+    /** Run time of the current sync (unit: milliseconds) */
     public long getRunTimeMs() {
         return runTimeMs;
     }
@@ -78,13 +105,16 @@ public final class SyncProgress {
     @Override
     public String toString() {
         return "SyncProgress{" +
-                "countLeft=" + countLeft +
-                "," + "countProcessed=" + countProcessed +
-                "," + "countTotal=" + countTotal +
-                "," + "countNew=" + countNew +
-                "," + "countNewUnread=" + countNewUnread +
-                "," + "countModified=" + countModified +
-                "," + "countDeleted=" + countDeleted +
+                "incomingMessagesProcessed=" + incomingMessagesProcessed +
+                "," + "incomingMessagesTotal=" + incomingMessagesTotal +
+                "," + "incomingMessagesNew=" + incomingMessagesNew +
+                "," + "incomingMessagesNewUnread=" + incomingMessagesNewUnread +
+                "," + "incomingMessagesModified=" + incomingMessagesModified +
+                "," + "incomingMessagesDeleted=" + incomingMessagesDeleted +
+                "," + "incomingAttachmentsDownloadedBytes=" + incomingAttachmentsDownloadedBytes +
+                "," + "incomingAttachmentsTotalBytes=" + incomingAttachmentsTotalBytes +
+                "," + "outgoingMessagesUploadedBytes=" + outgoingMessagesUploadedBytes +
+                "," + "outgoingMessagesTotalBytes=" + outgoingMessagesTotalBytes +
                 "," + "runTimeMs=" + runTimeMs +
         "}";
     }